July 2015 Update
Various small changes

Contact me via http://about.me/brian.aberle

+++ Servers/5LoavesDLL/5LoavesDLL.cpp
@@ -152,7 +152,7 @@ long pascal ServerStart()
 			s._len = wcslen(s._pWideStr);				
 #else
 			GString s(512);								
-			GetModuleFileName(0, s._str, s._len);		
+			GetModuleFileName(0, s._str, (unsigned long)s._len);		
 			s._len = strlen(s._str);						
 #endif
 			GString strStartupMessage;
